Refactor to add basic Android versions lambda
